Removed a bunch of things I was testing. Now, we finally have something on the right track when it comes to the event calendar.
I am simply using the javascript fullcalendar library, along with the google calendar API in order to render events from the google calendar onto my own custom made calendar. So far it is quite bare bones, but it is functioning as it should. There are some views and models that were created as a backup, in case I end up just using django models and views to create a calendar, and then use the admin panel to populate, but this seems cleaner so far.