Figured out how I am going to display so much text while keeping everything to only one page. Created modal text boxes that are currently opened with a button. There will be a small description and an image, with a "more info" button explaining the details.
The 'header' image on the left side works great, but every so often, with no changes at all, it will be too short for the container, so that a small bit shows at the bottom. It is fixed in position, but just something to figure out.

Still have bigger design issues to figure out, but much will be solved when the client gets back about font and colors.