Ready for preview.
The slider and title are adequate, just need to use Pillow or some other processor to alter images on the backend as they are uploaded. The slider is meant to be in landscape orientation, so portrait images that are uploaded will not look right.

The title is in a nice enough font, though the client will likely want to change it to their own. All in all. Good starting place.